fix : pycm_func.py splited
