+++ pycm/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from .pycm_func import *
+from .pycm_class_func import *
+from .pycm_overall_func import *
 from .pycm_output import *
 from .pycm_obj import *
 from .pycm_util import *
